pyproject.toml
[project]
name = "guardrail-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { email = "newton.pi.cheng@outlook.com" }
]
requires-python = "<=3.12, >=3.11"
dependencies = [
    "langchain-openai>=0.3.33",
    "langgraph>=0.6.7",
    "langgraph-api>=0.4.22",
    "langgraph-cli>=0.4.2",
    "pydantic>=2.11.9",
    "python-dotenv>=1.1.1",
]



[build-system]
requires = ["uv_build>=0.8.14,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "ruff>=0.13.1",
]

[project.scripts]
guardrail = "guardrail_service.cli:main"
